supercloud_console/requirements.txt
gradio>=4.0.0
requests
pandas
plotly
networkx
matplotlib
websocket-client
clickhouse-connect
python-dotenv
